refactor: ディレクトリ構造を見直した
